requirements.txt
streamlit
selenium
pandas
newspaper3k
bs4
cohere
tqdm